+++ packages/core/workers/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @atlaspack/workers
 
+## 2.14.42
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @atlaspack/logger@2.14.34
+  - @atlaspack/utils@3.2.3
+  - @atlaspack/types-internal@2.21.3
+  - @atlaspack/profiler@2.15.3
+
 ## 2.14.41
 
 ### Patch Changes