+++ packages/reporters/cli/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @atlaspack/reporter-cli
 
+## 2.17.24
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @atlaspack/utils@3.2.3
+  - @atlaspack/plugin@2.14.42
+  - @atlaspack/types@2.15.32
+
 ## 2.17.23
 
 ### Patch Changes