+++ packages/runtimes/service-worker/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @atlaspack/runtime-service-worker
 
+## 2.14.42
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @atlaspack/utils@3.2.3
+  - @atlaspack/plugin@2.14.42
+  - @atlaspack/types@2.15.32
+
 ## 2.14.41
 
 ### Patch Changes