+++ packages/transformers/toml/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @atlaspack/transformer-toml
 
+## 2.14.42
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @atlaspack/plugin@2.14.42
+
 ## 2.14.41
 
 ### Patch Changes